+++ src/js/PlotBoilerplate.js
@@ -56,8 +56,10 @@
  * @modified 2020-05-09 Included the Cirlcle class.
  * @modified 2020-06-22 Added the rasterScaleX and rasterScaleY config params.
  * @modified 2020-06-03 Fixed the selectedVerticesOnPolyon(Polygon) function: non-selectable vertices were selected too, before.
- * @modified 2020-06-06 Replacing Touchy.js by AlloyFinger.js
- * @version  1.8.2
+ * @modified 2020-07-06 Replacing Touchy.js by AlloyFinger.js
+ * @modified 2020-07-27 Added the getVertexNear(XYCoords,number) function
+ * @modified 2020-07-27 Extended the remove(Drawable) function: vertices are now removed, too.
+ * @version  1.9.0
  *
  * @file PlotBoilerplate
  * @fileoverview The main class.
@@ -538,15 +540,56 @@ var PlotBoilerplate = /** @class */ (function () {
      * @memberof PlotBoilerplate
      * @return {void}
      **/
-    PlotBoilerplate.prototype.remove = function (drawable, redraw) {
+    PlotBoilerplate.prototype.remove = function (drawable, redraw, removeWithVertices) {
         if (drawable instanceof Vertex_1.Vertex)
             this.removeVertex(drawable, false);
         for (var i = 0; i < this.drawables.length; i++) {
             if (this.drawables[i] === drawable) {
                 this.drawables.splice(i, 1);
-                // Check if some listeners need to be removed
-                if (drawable instanceof BezierPath_1.BezierPath)
-                    PlotBoilerplate.utils.disableBezierPathAutoAdjust(drawable);
+                if (removeWithVertices) {
+                    // Check if some listeners need to be removed
+                    if (drawable instanceof Line_1.Line) {
+                        // Add some lines
+                        this.removeVertex(drawable.a, false);
+                        this.removeVertex(drawable.b, false);
+                    }
+                    else if (drawable instanceof Vector_1.Vector) {
+                        this.removeVertex(drawable.a, false);
+                        this.removeVertex(drawable.b, false);
+                    }
+                    else if (drawable instanceof VEllipse_1.VEllipse) {
+                        this.removeVertex(drawable.center, false);
+                        this.removeVertex(drawable.axis, false);
+                    }
+                    else if (drawable instanceof Circle_1.Circle) {
+                        this.removeVertex(drawable.center, false);
+                    }
+                    else if (drawable instanceof Polygon_1.Polygon) {
+                        // for( var i in drawable.vertices )
+                        for (var i = 0; i < drawable.vertices.length; i++)
+                            this.removeVertex(drawable.vertices[i], false);
+                    }
+                    else if (drawable instanceof Triangle_1.Triangle) {
+                        this.removeVertex(drawable.a, false);
+                        this.removeVertex(drawable.b, false);
+                        this.removeVertex(drawable.c, false);
+                    }
+                    else if (drawable instanceof BezierPath_1.BezierPath) {
+                        PlotBoilerplate.utils.disableBezierPathAutoAdjust(drawable);
+                        for (var i = 0; i < drawable.bezierCurves.length; i++) {
+                            this.removeVertex(drawable.bezierCurves[i].startPoint, false);
+                            this.removeVertex(drawable.bezierCurves[i].startControlPoint, false);
+                            this.removeVertex(drawable.bezierCurves[i].endControlPoint, false);
+                            //if( i+1 == drawable.bezierCurves.length ) {
+                            this.removeVertex(drawable.bezierCurves[i].endPoint, false);
+                            //}
+                        }
+                    }
+                    else if (drawable instanceof PBImage_1.PBImage) {
+                        this.removeVertex(drawable.upperLeft, false);
+                        this.removeVertex(drawable.lowerRight, false);
+                    }
+                } // END removeWithVertices
                 if (redraw)
                     this.redraw();
                 return;
@@ -574,6 +617,18 @@ var PlotBoilerplate = /** @class */ (function () {
                 return;
             }
         }
+    };
+    ;
+    /**
+     * Find the vertex near the given position.
+     **/
+    PlotBoilerplate.prototype.getVertexNear = function (position, tolerance) {
+        var p = this.locatePointNear(position, // this.transformMousePosition(position.x, position.y),
+        tolerance / Math.min(this.config.cssScaleX, this.config.cssScaleY));
+        console.log(p);
+        if (p && p.typeName == "vertex")
+            return this.vertices[p.vindex];
+        return undefined;
     };
     ;
     /**
@@ -1343,41 +1398,11 @@ var PlotBoilerplate = /** @class */ (function () {
                 ;
             }
             else if (window["Touchy"] && typeof window["Touchy"] == "function") {
-                console.warn('(Deprecation) Found Touchy which support will stop soon. Please use AlloyFinger instead.');
+                console.error('[Deprecation] Found Touchy which is not supported any more. Please use AlloyFinger instead.');
                 // Convert absolute touch positions to relative DOM element position (relative to canvas)
-                // Some private vars to store the current mouse/position/button state.
-                var touchMovePos = null;
-                var touchDownPos = null;
-                var draggedElement = null;
-                var Touchy = (window["Touchy"]);
-                new Touchy(this.canvas, { one: function (hand, finger) {
-                        touchMovePos = new Vertex_1.Vertex(relPos_1(finger.lastPoint));
-                        touchDownPos = new Vertex_1.Vertex(relPos_1(finger.lastPoint));
-                        draggedElement = _self.locatePointNear(_self.transformMousePosition(touchMovePos.x, touchMovePos.y), PlotBoilerplate.DEFAULT_TOUCH_TOLERANCE / Math.min(_self.config.cssScaleX, _self.config.cssScaleY));
-                        if (draggedElement) {
-                            // The Touchy-points also have 'id' and 'time' attributes
-                            // which we are not interested in here.
-                            hand.on('move', function (points) {
-                                var rel = relPos_1(points[0]);
-                                var trans = _self.transformMousePosition(rel.x, rel.y);
-                                var diff = new Vertex_1.Vertex(_self.transformMousePosition(touchMovePos.x, touchMovePos.y)).difference(trans);
-                                if (draggedElement.typeName == 'vertex') {
-                                    if (!_self.vertices[draggedElement.vindex].attr.draggable)
-                                        return;
-                                    _self.vertices[draggedElement.vindex].add(diff);
-                                    var draggingVertex = _self.vertices[draggedElement.vindex];
-                                    var fakeEvent = { params: { dragAmount: diff.clone(), wasDragged: true, mouseDownPos: touchDownPos.clone(), mouseDragPos: touchDownPos.clone().add(diff), vertex: draggingVertex } };
-                                    draggingVertex.listeners.fireDragEvent(fakeEvent);
-                                    _self.redraw();
-                                }
-                                touchMovePos = new Vertex_1.Vertex(rel);
-                            });
-                        }
-                    }
-                });
             }
             else {
-                console.warn("Cannot initialize the touch handler. Touchy and AlloyFinger are missig. Did you include at least one of them?");
+                console.warn("Cannot initialize the touch handler. AlloyFinger is missig. Did you include it?");
             }
         }
         else {