+++ src/ts/BezierPath.ts
@@ -531,6 +531,18 @@ export class BezierPath implements SVGSerializable {
     };
     */
 
+    /*
+    insertVertexAt( t:number ) : void {
+	console.log('Inserting vertex at', t );
+	// Find the curve index
+	var u : number = 0;
+	var curveIndex : number = -1;
+	var localT : number = 0.0;
+	for( var i = 0; curveIndex == -1 && i < this.bezierCurves.length; i++ ) {
+	    
+	}
+    }; */
+
 
    
     /**