+++ changelog.md
@@ -1,6 +1,9 @@
 
 
 # Changelog
+* 2020-07-27
+  * Added the getVertexNear(XYCoords,number) function.
+  * Extended the remove(Drawable) function: vertices are now removed, too.
 * 2020-07-24 (1.8.2)
   * Fixed an error in the touch handlers 'tap start' function (nullpointerexception if empty area was tapped).
   * Added CubicBezierPath.getClosestT(Vertex) function.